requirements.txt
filetype==1.0.10
numpy==1.22.2
Pillow==9.0.1
